Create R. Permutation with arrays
Check if the elements of Array A are the same elements of Array B (the order is not important)
I solved this problem by summing the arrays and comparing the result. I tried to solve it by assigning a pointer and increasing it if an element of array A is similar to an element of Array b, after that, I compare the value of the pointer with the number of elements.